backend/requirements.txt
fastapi[all]==0.115.0
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
alembic==1.14.0
asyncpg==0.30.0
psycopg2-binary==2.9.10
redis==5.2.1
hiredis==3.0.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
pyotp==2.9.0
qrcode[pil]==7.4.2
email-validator==2.2.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-mock==3.14.0
httpx==0.28.1

# Utilities
python-dotenv==1.0.1
pydantic[email]==2.10.5
pydantic-settings==2.6.1
python-multipart==0.0.19
ulid-py==1.1.0

# Code Quality
black==24.10.0
flake8==7.1.1
mypy==1.14.0
isort==5.13.2
pre-commit==4.0.1

# Performance
orjson==3.10.12
msgpack==1.1.0